pyproject.toml
[project]
name = "coursework"
version = "0.1.0"
description = ""
authors = [
    {name = "Imanaliev Alisher",email = "imanalievalisher04@gmail.com"}
]

requires-python = ">=3.10,<4"
dependencies = [
    "amqp (==5.3.1)",
    "anyio (==4.9.0)",
    "asgiref (==3.8.1)",
    "argon2-cffi (==23.1.0)",
    "billiard (==4.2.1)",
    "djangorestframework (==3.15.2)",
    "djangorestframework-simplejwt (==5.4.0)",
    "pyjwt (==2.10.1)",
    "django-cors-headers (==4.7.0)",
    "drf-spectacular (==0.28.0)",
    "build (==1.2.2.post1)",
    "cachecontrol (==0.14.3)",
    "celery (==5.5.2)",
    "certifi (==2025.4.26)",
    "charset-normalizer (==3.4.2)",
    "cleo (==2.1.0)",
    "click (==8.1.8)",
    "click-didyoumean (==0.3.1)",
    "click-plugins (==1.1.1)",
    "click-repl (==0.3.0)",
    "colorama (==0.4.6)",
    "crashtest (==0.4.1)",
    "crispy-bootstrap5 (==2025.4)",
    "cron-descriptor (==1.4.5)",
    "distlib (==0.3.9)",
    "django (==5.2)",
    "django-anymail (==13.0)",
    "django-celery-beat (==2.8.0)",
    "django-countries (==7.6.1)",
    "django-crispy-forms (==2.4)",
    "django-debug-toolbar (==5.2.0)",
    "django-enumfields (==2.1.1)",
    "django-environ (==0.12.0)",
    "django-extensions (==4.1)",
    "django-jazzmin (==3.0.1)",
    "django-model-utils (==5.0.0)",
    "django-redis (==5.4.0)",
    "django-timezone-field (==7.1)",
    "dulwich (==0.22.8)",
    "fastjsonschema (==2.21.1)",
    "filelock (==3.18.0)",
    "findpython (==0.6.3)",
    "gunicorn (==23.0.0)",
    "h11 (==0.16.0)",
    "hiredis (==3.1.0)",
    "httpcore (==1.0.9)",
    "httpx (==0.28.1)",
    "idna (==3.10)",
    "installer (==0.7.0)",
    "jaraco-classes (==3.4.0)",
    "jaraco-context (==6.0.1)",
    "jaraco-functools (==4.1.0)",
    "keyring (==25.6.0)",
    "kombu (==5.5.3)",
    "more-itertools (==10.7.0)",
    "msgpack (==1.1.0)",
    "packaging (==25.0)",
    "pbs-installer (==2025.4.9)",
    "pkginfo (==1.12.1.2)",
    "platformdirs (==4.3.7)",
    "poetry (==2.1.2)",
    "poetry-core (==2.1.2)",
    "prompt-toolkit (==3.0.51)",
    "psycopg2-binary (==2.9.10)",
    "psycopg2 (==2.9.10)",
    "pyproject-hooks (==1.2.0)",
    "python-crontab (==3.2.0)",
    "python-dateutil (==2.9.0.post0)",
    "pywin32-ctypes (==0.2.3)",
    "rapidfuzz (==3.13.0)",
    "redis (==6.0.0)",
    "reportlab (>=4.0.0,<5.0.0)",
    "openpyxl (>=3.1.0,<4.0.0)",
    "pytest (>=8.0.0,<9.0.0)",
    "pytest-django (>=4.8.0,<5.0.0)",
    "pytest-cov (>=4.1.0,<5.0.0)",
    "factory-boy (>=3.3.0,<4.0.0)",
    "faker (>=24.0.0,<25.0.0)",
    "requests (==2.32.3)",
    "requests-toolbelt (==1.0.0)",
    "shellingham (==1.5.4)",
    "six (==1.17.0)",
    "sniffio (==1.3.1)",
    "sqlparse (==0.5.3)",
    "tomlkit (==0.13.2)",
    "trove-classifiers (==2025.5.1.12)",
    "typing-extensions (==4.13.2)",
    "tzdata (==2025.2)",
    "urllib3 (==2.4.0)",
    "uvicorn (==0.34.2)",
    "uvicorn-worker (==0.3.0)",
    "vine (==5.1.0)",
    "virtualenv (==20.30.0)",
    "watchfiles (==1.0.5)",
    "wcwidth (==0.2.13)",
    "zstandard (==0.23.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
